- Ny DTO (Oppgave) og litt restrukturering som følge. - Legger til eksplisit mapping til/fra json (usikker på om det håndteres automatisk eller ikke)
