+++ src/main/kotlin/no/nav/familie/ks/kontrakter/Kontrakt.kt
@@ -1,0 +1,11 @@
+package no.nav.familie.ks.kontrakter
+
+import com.fasterxml.jackson.databind.DeserializationFeature
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.module.kotlin.registerKotlinModule
+
+open class Kontrakt {
+    protected val objectMapper = ObjectMapper()
+        .registerKotlinModule()
+        .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
+}
