+++ src/main/kotlin/no/nav/familie/ks/kontrakter/søknad/OppgittErklæring.kt
@@ -1,4 +1,4 @@
-package no.nav.familie.ks.søknadkontrakt
+package no.nav.familie.ks.kontrakter.søknad
 
 data class OppgittErklæring(
         val isBarnetHjemmeværendeOgIkkeAdoptert: Boolean,
