+++ src/main/kotlin/no/nav/familie/ks/kontrakter/søknad/OppgittFamilieforhold.kt
@@ -1,4 +1,4 @@
-package no.nav.familie.ks.søknadkontrakt
+package no.nav.familie.ks.kontrakter.søknad
 
 data class OppgittFamilieforhold(
     val barna: Set<Barn>,
