+++ src/main/kotlin/no/nav/familie/ks/kontrakter/søknad/Standpunkt.kt
@@ -1,4 +1,4 @@
-package no.nav.familie.ks.søknadkontrakt
+package no.nav.familie.ks.kontrakter.søknad
 
 enum class Standpunkt {
     JA, NEI, UBESVART
