+++ src/main/kotlin/no/nav/familie/ks/kontrakter/søknad/Søknad.kt
@@ -1,0 +1,18 @@
+package no.nav.familie.ks.kontrakter.søknad
+
+import com.fasterxml.jackson.module.kotlin.readValue
+import no.nav.familie.ks.kontrakter.Kontrakt
+import java.time.LocalDateTime
+import javax.validation.constraints.Pattern
+
+data class Søknad(
+    @Pattern(regexp = "[0-9]{11}")
+    val søkerFnr: String,
+    val innsendtTidspunkt: LocalDateTime,
+    val oppgittFamilieforhold: OppgittFamilieforhold,
+    val oppgittUtlandsTilknytning: OppgittUtlandsTilknytning,
+    val oppgittErklæring: OppgittErklæring
+) : Kontrakt() {
+    fun Søknad.toJson(): String = objectMapper.writeValueAsString(this)
+    fun String.toSøknad(): Søknad = objectMapper.readValue(this)
+}
