--- src/main/kotlin/no/nav/familie/ks/søknadkontrakt/Søknad.kt
@@ -1,13 +1,0 @@
-package no.nav.familie.ks.søknadkontrakt
-
-import java.time.LocalDateTime
-import javax.validation.constraints.Pattern
-
-data class Søknad(
-        @Pattern(regexp = "[0-9]{11}")
-        val søkerFnr: String,
-        val innsendtTidspunkt: LocalDateTime,
-        val oppgittFamilieforhold: OppgittFamilieforhold,
-        val oppgittUtlandsTilknytning: OppgittUtlandsTilknytning,
-        val oppgittErklæring: OppgittErklæring
-)
